propensityForest {causalTree}	R Documentation

Propensity Forest for Causal Effect Regression and Estimation (Modified Causal Tree Ensembles)

Description

Fit and evaluate a user selected number of causalTree models to get an ensemble of rpart objects Trees are split using covariates and the treatment vector instead of the outcome variable, and evaluated using the complete data covariates and the actual outcome variable.

Usage

propensityForest(formula, data, treatment,  
                         na.action, 
                         split.Rule, split.Honest, split.Bucket, bucketNum,
                         bucketMax, cv.option, cv.Honest, minsize, 
                         propensity, control, split.alpha, cv.alpha,  
                         
                         sample.size.total, sample.size.train.frac = 1,
                         mtry, nodesize, num.trees,ncolx,ncov_sample)

Arguments

formula
a formula, with a response and features but no interaction terms. If this a a data frome, that is taken as the model frame (see model.frame).

data
an optional data frame that includes the variables named in the formula.

weights
optional case weights.

treatment
a vector that indicates the treatment status of each observation. 1 represents treated and 0 represents control. Only binary treatment supported in this version.

na.action
the default action deletes all observations for which y is missing, but keeps those in which one or more predictors are missing.

split.Rule
causalTree splitting options, one of "TOT", "CT", "fit", "tstats", four splitting rules in causalTree. Note that the "tstats" alternative does not have an associated cross-validation method cv.option; see Athey and Imbens (2016) for a discussion. Note further that split.Rule and cv.option can mix and match.

split.Honest
boolean option, TRUE or FALSE, used for split.Rule as "CT" or "fit". If set as TRUE, do honest splitting, with default split.alpha = 0.5; if set as FALSE, do adaptive splitting with split.alpha = 1. The user choice of split.alpha will be ignored if split.Honest is set as FALSE, but will be respected if set to TRUE. For split.Rule="TOT", there is no honest splitting option and the parameter split.alpha does not matter. For split.Rule="tstats", a value of TRUE enables use of split.alpha in calculating the risk function, which determines the order of pruning in cross-validation. Note also that causalTree function returns the estimates from the training data, no matter what the value of split.Honest is; the tree must be re-estimated to get the honest estimates using estimate.causalTree. The wrapper function honest.CausalTree does honest estimation in one step and returns a tree.

split.Bucket
boolean option, TRUE or FALSE, used to specify whether to apply the discrete method in splitting the tree. If set as TRUE, in splitting a node, the observations in a leaf will be be partitioned into buckets, with each bucket containing bucketNum treated and bucketNum control units, and where observations are ordered prior to partitioning. Splitting will take place by bucket.

bucketNum
number of observations in each bucket when set split.Bucket = TRUE. However, the code will override this choice in order to guarantee that there are at least minsize and at most bucketMax buckets.

bucketMax
Option to choose maximum number of buckets to use in splitting when set split.Bucket = TRUE, bucketNum can change by choice of bucketMax.

cv.option
cross validation options, one of "TOT", "matching", "CT", "fit", four cross validation methods in causalTree. There is no cv.option for the split.Rule "tstats"; see Athey and Imbens (2016) for discussion.

cv.Honest
boolean option, TRUE or FALSE, only used for cv.option as "CT" or "fit", to specify whether to apply honest risk evalation function in cross validation. If set TRUE, use honest risk function, otherwise use adaptive risk function in cross validation. If set FALSE, the user choice of cv.alpha will be set to 1. If set TRUE, cv.alpha will default to 0.5, but the user choice of cv.alpha will be respected. Note that honest cv estimates within-leaf variances and may perform better with larger leaf sizes and/or small number of cross-validation sets.

minsize
in order to split, each leaf must have at least minsize treated cases and minsize control cases. The default value is set as 2.

propensity
propensity score used in "TOT" splitting and "TOT", honest "CT" cross validation methods. The default value is the proportion of treated cases in all observations. In this implementation, the propensity score is a constant for the whole dataset. Unit-specific propensity scores are not supported; however, the user may use inverse propensity scores as case weights if desired.

control
a list of options that control details of the rpart algorithm. See rpart.control.

split.alpha
scale parameter between 0 and 1, used in splitting risk evaluation function for "CT". When split.Honest = FALSE, split.alpha will be set as 1. For split.Rule="tstats", if split.Honest=TRUE, split.alpha is used in calculating the risk function, which determines the order of pruning in cross-validation.

cv.alpha
scale paramter between 0 and 1, used in cross validation risk evaluation function for "CT" and "fit". When cv.Honest = FALSE, cv.alpha will be set as 1.

cost
a vector of non-negative costs, one for each variable in the model. Defaults to one for all variables. These are scalings to be applied when considering splits, so the improvement on splitting on a variable is divided by its cost in deciding which split to choose.

sample.size.total
Sample size used to build each tree in the forest (sampled randomly with replacement)

sample.size.train.frac
Fraction of the sample size used for building each tree (training)

mtry
Number of data features used to build a tree (This variable is not used presently)

nodesize
Minimum number of observations for treated and control cases in one leaf node

num.trees
Number of trees to be built in the causal forest

ncolx
Total number of covariates

ncov_sample
Number of covariates randomly sampled to build each tree in the forest

...
arguments to rpart.control may also be specified in the call to causalForest. They are checked against the list of valid arguments. An example of a commonly set parameter would be xval, which sets the number of cross-validation samples. The parameter minsize is implemented differently in causalTree than in rpart; we require a minimum of minsize treated observations and a minimum of minsize control observations in each leaf.

Details

Propensity forest is similar to a causal forest, with some important differences as discussed below. The causalForest builds an ensemble of CausalTrees, by repeated random sampling of the data with replacement. For prediction, the average value over all tree predictions is used. Propensity forest differs from a causal forest in the following way: The tree building phase is done by using the covariates and treatment vector as the dummy output/outcomes variable (instead of the actual outcomes variable) During the tree evaluation phase, the reestimation error is calculated on the actual outcomes variable to evaluate the tree performance. Note that the propensity forest will always build an adaptive (non honest) ensemble of trees.

CausalTree differs from rpart function from rpart package in splitting rules and cross validation methods. Please check Athey and Imbens, Recursive Partitioning for Heterogeneous Causal Effects (2016) and Stefan Wager and Susan Athey, Estimation and Inference of Heterogeneous Treatment Effects using Random Forests for more details.

Value

An object of class rpart. See rpart.object.

References

Breiman L., Friedman J. H., Olshen R. A., and Stone, C. J. (1984) Classification and Regression Trees. Wadsworth.

Athey, S and G Imbens (2016) Recursive Partitioning for Heterogeneous Causal Effects. http://arxiv.org/abs/1504.01132

Wager,S and Athey, S (2015) Estimation and Inference of Heterogeneous Treatment Effects using Random Forests http://arxiv.org/abs/1510.04342

See Also

honest.causalTree, rpart.control, rpart.object, summary.rpart, rpart.plot

Examples

pf <- propensityForest(as.formula(paste("y~",f)), data=dataTrain, treatment=dataTrain$w, 
                   split.Bucket=F, 
                   sample.size.total = floor(nrow(dataTrain) / 2), 
                   mtry = ceiling(ncol(dataTrain)/3), nodesize = 25, num.trees=num.trees.temp,ncolx=ncolx,ncov_sample=ncov_sample) 

pfpredtest <- predict(pf, newdata=dataTest, type="vector")
plot(dataTest$tau_true,pfpredtest)

pfpredtrainall <- predict(pf, newdata=dataTrain, predict.all = TRUE, type="vector")

[Package causalTree version 0.0 Index]