simulation.2 {causalTree}	R Documentation

simulation data from Susan and Guido's paper

Description

The simulation.2 data frame has 500 rows and 12 columns, giving 10 x features, corresponding y value, and treatment status for each 500 samples.

Usage

simulation.2

Format

This data frame contains the following columns:

x1,...,x10

numeric vectors giving the 10 feature values, each follows standard normal distribution.

y

numeric vectors giving the response, generated from normal distribution depending on x.

treatment

treatment status vector, 1 represents treated and 0 represents controlled.

Source

One simulation case in Susan and Guido's paper, similar as simulation.1.

See Also

simulation.1

[Package causalTree version 0.0 Index]